feat: move classes to their own files
