+++ oracle-api/src/main/java/ca/bc/gov/oracleapi/dto/consep/Numbers.java
@@ -1,0 +1,15 @@
+package ca.bc.gov.oracleapi.dto.consep;
+
+import java.util.List;
+
+public class Numbers {
+  private List<Double> numbers;
+
+  public List<Double> getNumbers() {
+      return numbers;
+  }
+
+  public void setNumbers(List<Double> numbers) {
+      this.numbers = numbers;
+  }
+}
